Organized codebase so that everything is in the correct folders and named in a way to be able to find it easier.
