+++ covid-19-app/.firebase/hosting.YnVpbGQ.cache
@@ -3,23 +3,23 @@ logo192.png,1585071846145,3ee59515172ee198f3be375979df15ac5345183e656720a381b887
 logo512.png,1585071846145,ee7e2f3fdb8209c4b6fd7bef6ba50d1b9dba30a25bb5c3126df057e1cb6f5331
 manifest.json,1585071846145,aff3449bdc238776f5d6d967f19ec491b36aed5fb7f23ccff6500736fd58494a
 robots.txt,1585071846145,bfe106a3fb878dc83461c86818bf74fc1bdc7f28538ba613cd3e775516ce8b49
-index.html,1619733184261,2dd0d20e955420111a6b409bd7bc3d6f7924a9d04756ced1461b7cc924ab786b
-precache-manifest.b63604e6bd8822f88c5f3decdeb67719.js,1619733184262,97c4f07336b2c27da2dc790aa1d8d4db4d93dd4bc38d61250f2e3f435e7319a9
-service-worker.js,1619733184262,9349b3e382a6fe26054deb20a279e82e2eb55b40342689164c0c19fab03f3cb2
-asset-manifest.json,1619733184262,924719994be6fe2d18e463d5662b81da905cb1080174ad2a25cccb6b921d68f8
-static/css/main.49fd0884.chunk.css,1619733184232,2a3565410a9c63231a275d52d23eb3503089441ffb746eaf33f406e7513328d3
-static/css/main.49fd0884.chunk.css.map,1619733184233,08792b331dd9fbdade0a6cb022bddd1c63f0a4d4b98ac253c17612c4559fe54f
-static/css/2.6aa45fda.chunk.css,1619733184233,2f3389356096d567dbfa2be02c34466bf9d5a5ea78127cb17ce74f6340e9faa9
-static/js/main.df4ebe3b.chunk.js,1619733184233,c6fdd3797e24a7199a3124a1b81e0b39bcd33ba099a047f8645b67197986e1ef
-static/js/2.b1f228eb.chunk.js.LICENSE.txt,1619733184233,a38291cb5134f88936364fb5110fa6e44d470615076fbf89f2fa980be0480370
-static/js/runtime-main.bd5efe05.js,1619733184233,a2d806a9a9f209afaebf82a0dd4e76b583aa1cb754b1a473157a695a0e8c1b99
-static/js/runtime-main.bd5efe05.js.map,1619733184229,236a985fc48ee914288773b6553dd4a34df42e30da4450a5a3893ffad934c574
-static/css/2.6aa45fda.chunk.css.map,1619733184233,5f8b2210590c0f57f67e017769afbe5f2cce1a17a3fcbeffdb76038dca99e728
-static/js/main.df4ebe3b.chunk.js.map,1619733184233,e7d0729af239d7afe2e03d7c933f3b0f43af46146ccfa65719c9ad330662baff
-static/media/weathericons-regular-webfont.1cd48d78.woff2,1619733184233,d87a1b46c23762fb8c69f216b841ebb9fdd33a3aab719f3cc0a441f61d85942c
-static/media/weathericons-regular-webfont.8cac70eb.woff,1619733184229,722602fd6e482bffb4b06c7df54574ade45cfb6faebf26dc8a97ab38ff7efe29
-static/media/weathericons-regular-webfont.4b658767.eot,1619733184233,423a6f6caee8393f1f6dd7db1aad3a0e7bae3c48e1cee4932c91a30cc978ff60
-static/media/weathericons-regular-webfont.4618f0de.ttf,1619733184233,ba39ae184819f2fd3f29016f3ca2f574a8254e39c809b8f5e200137518a1120e
-static/media/weathericons-regular-webfont.ed0943f2.svg,1619733184233,790d040314f25278dc4a38fc17ee5f0d395ee2c5477063b3cbfa2f74e8d330fb
-static/js/2.b1f228eb.chunk.js,1619733184234,378dca28326ac241e3e2b132d48c692821edcaed4cde101525320d900ae3939d
-static/js/2.b1f228eb.chunk.js.map,1619733184264,b1537d8af4956daa9c6cfeb81ffc8d2f0d70f61e9702c04682eb836dfc685741
+asset-manifest.json,1619737700302,4db935c5c3f7f1c00195cc27a08f84a1e4b8b20eeebc569d8e9b33494b84d7f5
+index.html,1619737700302,82d1b5c52a316e3dcbf03311faf4fd3224ae7636bbf214007d2d46bb4252a607
+static/css/main.49fd0884.chunk.css,1619737700271,2a3565410a9c63231a275d52d23eb3503089441ffb746eaf33f406e7513328d3
+static/css/main.49fd0884.chunk.css.map,1619737700271,08792b331dd9fbdade0a6cb022bddd1c63f0a4d4b98ac253c17612c4559fe54f
+precache-manifest.6eb8e2ecf1cb4008368a6d0fe30117e9.js,1619737700302,2d7aa032fb4e1393c5209786223c8b9198242c571597437daa7e336bb55959c6
+static/js/2.a8564617.chunk.js.LICENSE.txt,1619737700271,a38291cb5134f88936364fb5110fa6e44d470615076fbf89f2fa980be0480370
+service-worker.js,1619737700302,0b0614f0a2901a7276a4e7122343447f826a9f04734c5f102a6071d5fd632f2b
+static/css/2.6aa45fda.chunk.css,1619737700271,2f3389356096d567dbfa2be02c34466bf9d5a5ea78127cb17ce74f6340e9faa9
+static/js/main.a1ee5428.chunk.js,1619737700271,80d4cdb33afd19abf765e1983836fb1d1773b27c203f09e288f942460a7294e9
+static/js/runtime-main.bd5efe05.js,1619737700271,a2d806a9a9f209afaebf82a0dd4e76b583aa1cb754b1a473157a695a0e8c1b99
+static/js/runtime-main.bd5efe05.js.map,1619737700271,236a985fc48ee914288773b6553dd4a34df42e30da4450a5a3893ffad934c574
+static/css/2.6aa45fda.chunk.css.map,1619737700271,5f8b2210590c0f57f67e017769afbe5f2cce1a17a3fcbeffdb76038dca99e728
+static/js/main.a1ee5428.chunk.js.map,1619737700271,f0bd2aa70808b2aadf57fcafef97c2ded9c5ef0b751905d2cb6a3c19c75880b5
+static/media/weathericons-regular-webfont.1cd48d78.woff2,1619737700271,d87a1b46c23762fb8c69f216b841ebb9fdd33a3aab719f3cc0a441f61d85942c
+static/media/weathericons-regular-webfont.8cac70eb.woff,1619737700271,722602fd6e482bffb4b06c7df54574ade45cfb6faebf26dc8a97ab38ff7efe29
+static/media/weathericons-regular-webfont.4618f0de.ttf,1619737700271,ba39ae184819f2fd3f29016f3ca2f574a8254e39c809b8f5e200137518a1120e
+static/media/weathericons-regular-webfont.4b658767.eot,1619737700271,423a6f6caee8393f1f6dd7db1aad3a0e7bae3c48e1cee4932c91a30cc978ff60
+static/media/weathericons-regular-webfont.ed0943f2.svg,1619737700271,790d040314f25278dc4a38fc17ee5f0d395ee2c5477063b3cbfa2f74e8d330fb
+static/js/2.a8564617.chunk.js,1619737700271,06b67eec6fd4fabeac883113b0f981f8aadf0251b10c14c71ab132f2fa92a470
+static/js/2.a8564617.chunk.js.map,1619737700302,5dbb723c9187763aab26b9b209a67d0e51eb2de995e56ab8555d965debbc9296
