+++ covid-19-app/src/CSS/InfoBox.css
