+++ covid-19-app/src/CSS/Table.css
