+++ covid-19-app/src/components/InfoBox.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './InfoBox.css';
+import '../CSS/InfoBox.css';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import Typography from '@material-ui/core/Typography';
