+++ covid-19-app/src/components/LineGraph.js
