+++ covid-19-app/src/components/Map.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import './Map.css';
+import '../CSS/Map.css';
 import { Map as LeafletMap, TileLayer } from 'react-leaflet';
 import { showDataOnMap } from './util';
