+++ covid-19-app/src/components/PrivateRoute.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Route, Redirect } from "react-router-dom";
-import { useUser } from "./hooks";
+import { useUser } from '../hooks';
 
 const PrivateRoute = ({ component: Component, ...rest }) => {
   const currentUser = useUser();
