+++ covid-19-app/src/components/Table.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Table.css';
+import '../CSS/Table.css';
 import numeral from "numeral";
 
 function Table({countries}) {
