+++ covid-19-app/src/components/util.js
