+++ covid-19-app/src/config/fire.js
