+++ covid-19-app/src/providers/AuthProvider.js
@@ -1,6 +1,6 @@
   
 import React, { useState, useEffect } from "react";
-import fire from "../fire";
+import fire from "../config/fire";
 
 export const AuthContext = React.createContext();
